requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.30.0
bcrypt==5.0.0
black==26.3.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.6
greenlet==3.1.1
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
mypy_extensions==1.1.0
numpy==2.4.3
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.4
preshed==3.0.12
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.7.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.20
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.2
starlette==0.41.3
thinc==8.3.10
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==16.0
wrapt==2.1.2
cloudinary==1.38.0
google-genai==1.3.0
